Move utility and domain functions out of client.py
